Free merged child from memory.
While I'm at it, delete two wrong tests in bpt_delete() and bpt_insert().
This is because the first leaf node passed to the internal functions
(i.e. bpt_delete_internal and bpt_insert_internal) can become null
or invalid, by the effect of merge or split. At this point, I didn't
come up with a short and quick alternative for the tests, so I removed
them.